# _docs/openapi/cross-stack-measure-report-exchange.openapi.yaml
openapi: 3.0.3
info:
  title: Cross-Stack DEQM MeasureReport Exchange
  version: 1.0.0
  description: |
    OpenAPI contract for the bidirectional DEQM MeasureReport exchange between
    the **submitters** stack (producer) and the **receivers** stack (consumer).

    ## Producer (submitters)

    - `POST /api/workbench/cohorts/{cohortId}/measure-reports?reportType=...`
      builds the cohort report in the chosen DEQM profile and dispatches it to the
      receivers `/api/workbench/measure-reports` route.

    - `POST /api/workbench/cohorts/{cohortId}/measure-summary/send` (legacy)
      continues to dispatch the proprietary `measure_summary` rollup to
      `/api/workbench/measure-summaries` for back-compat. **Deprecated.**

    ## Consumer (receivers)

    - `POST /api/workbench/measure-reports` — accepts a DEQM MeasureReport or a
      Bundle of MeasureReports, validates DEQM mandatory elements (deqm-0, deqm-1),
      and persists by `type` under the appropriate `docType`.

    - `POST /api/workbench/measure-summaries` (legacy) — still accepts the
      proprietary cohort rollup for one release. **Deprecated.**

    ## reportType discriminator

    The `reportType` field on the producer drives which FHIR payload is built:

    | `reportType`   | FHIR `MeasureReport.type` | Cosmos `docType`              |
    |----------------|---------------------------|-------------------------------|
    | `individual`   | `individual`              | `measure_report`              |
...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/workbench/measure-reports:
    post:
      summary: Accept and persist a DEQM MeasureReport or Bundle
      operationId: receiveMeasureReport
      description: |
        Accepts a single DEQM MeasureReport or a Bundle of MeasureReports.

        - Validates DEQM mandatory elements (deqm-0, deqm-1) on each report.
        - Persists each report using a `docType` derived from `MeasureReport.type`.
        - Returns an `OperationOutcome` (400) for any conformance violation.
      tags:
        - consumer
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/SummaryMeasureReport'
                - $ref: '#/components/schemas/SubjectListMeasureReport'
                - $ref: '#/components/schemas/IndividualMeasureReport'
                - $ref: '#/components/schemas/MeasureReportBundle'
              discriminator:
                propertyName: resourceType
            examples:
              summaryIngest:
                summary: Ingest a Summary MeasureReport
                value:
                  resourceType: MeasureReport
                  id: report-sum-001
                  status: complete
                  type: summary
                  measure: 'https://example.org/Measure/CMS165v9|9.0.000'
                  reporter:
                    reference: Organization/submitters-org
                  date: '2025-12-31T23:59:59Z'
                  period:
                    start: '2025-01-01'
                    end: '2025-12-31'
                  group:
                    - population:
                        - code:
                            coding:
                              - system: 'http://terminology.hl7.org/CodeSystem/measure-population'
                                code: denominator
                          count: 8
                        - code:
                            coding:
                              - system: 'http://terminology.hl7.org/CodeSystem/measure-population'
                                code: numerator
                          count: 6
                      measureScore:
                        value: 0.75
      responses:
        '200':
          description: Report(s) accepted and persisted.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SingleReportResponse'
                  - $ref: '#/components/schemas/BundleIngestResponse'
        '400':
          description: DEQM conformance violation or unsupported resourceType.
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '401':
          description: Authentication required.
        '500':
          description: Persistence failure.
    get:
      summary: List persisted DEQM MeasureReports
      operationId: listMeasureReports
      tags:
        - consumer
      parameters:
        - name: reportType
          in: query
          description: Filter by DEQM profile type.
          schema:
            type: string
            enum: [individual, subject-list, summary]
      responses:
        '200':
          description: List of persisted MeasureReport index documents.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasureReportListResponse'
...
  /api/workbench/measure-summaries:
    post:
      summary: (Deprecated) Accept a proprietary cohort measure summary
      operationId: receiveMeasureSummaryLegacy
      deprecated: true
      description: |
        **Deprecated.** Use `POST /api/workbench/measure-reports` instead.

        Accepts the proprietary `measure_summary` rollup from submitters and persists
        it as `docType=measure_summary`. Kept for one release for back-compat.
      tags:
        - consumer
        - deprecated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasureSummaryPayload'
      responses:
        '200':
          description: Summary accepted and persisted.
        '400':
          description: Request body failed validation.
        '500':
          description: Persistence failure.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    FHIRReference:
      type: object
      properties:
        reference:
          type: string
        display:
          type: string
    MeasurePopulationBase:
      type: object
      required: [code, count]
      properties:
        code:
          type: object
          properties:
            coding:
              type: array
              items:
                type: object
                properties:
                  system:
                    type: string
                  code:
                    type: string
                    enum: [initial-population, denominator, numerator, denominator-exclusion]
        count:
          type: integer
          minimum: 0
    IndividualMeasureReport:
      type: object
      required: [resourceType, status, type, measure, reporter, date, period]
      properties:
        resourceType:
          type: string
          enum: [MeasureReport]
        id:
          type: string
        status:
          type: string
          enum: [complete, pending, error]
        type:
          type: string
          enum: [individual]
        measure:
          type: string
          description: 'Canonical URL including version, e.g. .../Measure/CMS165v9|9.0.000'
        subject:
          $ref: '#/components/schemas/FHIRReference'
        reporter:
          $ref: '#/components/schemas/FHIRReference'
        date:
          type: string
          format: date-time
        period:
          type: object
          required: [start, end]
          properties:
            start:
              type: string
              format: date
            end:
              type: string
              format: date
        group:
          type: array
          items:
            type: object
            properties:
              population:
                type: array
                items:
                  $ref: '#/components/schemas/MeasurePopulationBase'
    SubjectListMeasureReport:
      allOf:
        - $ref: '#/components/schemas/IndividualMeasureReport'
        - type: object
          required: [group]
          properties:
            type:
              type: string
              enum: [subject-list]
            group:
              type: array
              minItems: 1
              items:
                type: object
                properties:
                  population:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/MeasurePopulationBase'
                        - type: object
                          properties:
                            subjectResults:
                              $ref: '#/components/schemas/FHIRReference'
            contained:
              type: array
              items:
                type: object
    SummaryMeasureReport:
      allOf:
        - $ref: '#/components/schemas/IndividualMeasureReport'
        - type: object
          required: [group]
          properties:
            type:
              type: string
              enum: [summary]
            group:
              type: array
              minItems: 1
              items:
                type: object
                properties:
                  population:
                    type: array
                    items:
                      $ref: '#/components/schemas/MeasurePopulationBase'
                  measureScore:
                    type: object
                    properties:
                      value:
                        type: number
                        format: double
    MeasureReportBundle:
      type: object
      required: [resourceType, type, entry]
      properties:
        resourceType:
          type: string
          enum: [Bundle]
        id:
          type: string
        type:
          type: string
          enum: [collection, transaction, batch]
        entry:
          type: array
          items:
            type: object
            properties:
              resource:
                oneOf:
                  - $ref: '#/components/schemas/IndividualMeasureReport'
                  - $ref: '#/components/schemas/SubjectListMeasureReport'
                  - $ref: '#/components/schemas/SummaryMeasureReport'
...
    SingleReportResponse:
      type: object
      properties:
        report:
          $ref: '#/components/schemas/PersistedMeasureReport'
    BundleIngestResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: '#/components/schemas/PersistedMeasureReport'
        count:
          type: integer
    MeasureReportListResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: '#/components/schemas/PersistedMeasureReport'
        count:
          type: integer
...
    MeasureSummaryPayload:
      type: object
      description: Legacy proprietary cohort rollup payload (deprecated).
      required: [agency, cohort]
      properties:
        id:
          type: string
          nullable: true
        sourceStack:
          type: string
          default: submitters
        agency:
          type: object
          required: [id]
          properties:
            id:
              type: string
        cohort:
          type: object
          required: [id]
          properties:
            id:
              type: string
        measureIds:
          type: array
          items:
            type: string
    OperationOutcome:
      type: object
      required: [resourceType, issue]
      properties:
        resourceType:
          type: string
          enum: [OperationOutcome]
        issue:
          type: array
          items:
            type: object
            required: [severity, code]
            properties:
              severity:
                type: string
                enum: [fatal, error, warning, information]
              code:
                type: string
              diagnostics:
                type: string
...